+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+* Do not throw error on left click in chat if no Kingdom Sheet actor is present
+
 ## [0.0.19] - 2023-02-20
 
 ### Fixed